requirements.txt
pandas
numpy
scikit-learn
category_encoders
seaborn
matplotlib
pyyaml
lightgbm
tabulate
imbalanced-learn
geocoder
optuna
xgboost
japanize_matplotlib
